Full implimentation of network "layer" in epstein
It's a "layer" in the sense that it's an interaction layer, but it's not
implimented on a board, instead as a list that agents interact with
regard to spatial distance.

The layer is constructed with preference to location at intitialization
for possible future improvements, but since agents move randomly it
makes no difference.